Crate rsp2_tasks_config

Crate where serde_yaml code for the ‘tasks’ crate is monomorphized, because this is a huge compile time sink.

The functions here also make use of serde_ignored to catch typos in the config.

Structs

AcousticSearch

AlwaysFail

Animate
Output normal modes each iteration of the ev loop, for visualization purposes.

Cg

DeprecatedLammpsSettings

EnergyPlotSettings

EvLoop
Options describing the ev-loop.

Lammps

LammpsPotentialAirebo

LammpsPotentialKolmogorovCrespiFull

LammpsPotentialKolmogorovCrespiZ

LammpsPotentialRebo

LatticeRelax

LayerSearch

MaskBit

Masses
Masses by element.

MessagePhononEigensolverPhonopy

OldPotentialKolmogorovCrespiZ

Phonons

Potential

PotentialDftbPlus

PotentialKolmogorovCrespi

PotentialReboNonreactive

ScaleRanges

Settings
Raw deserialized form of settings.

Snapshot

ValidatedEnergyPlotSettings

ValidatedPotential
Potential settings known to satisfy certain properties:

ValidatedSettings
Root settings object.

Enums

AnimateFormat

AnimateWhich

CgFlavor

CgOnLsFailure
Behavior when a linesearch along the steepest descent direction fails. (this is phenomenally rare for the Hager linesearch method, and when it does occur it may very well be due to exceptionally good convergence, rather than any sort of actual failure)

EigenvectorChase

KolmogorovCrespiNormals

KolmogorovCrespiParams

LammpsKolmogorovCrespiParams

LammpsPotentialKind

LammpsUpdateStyle

Never

Parameter

PhononDispFinder

PhononDispFinderRsp2Directions

PhononEigensolver

PhononSumRule

PotentialKind

PotentialReboNewParams

Scalable

ScalableRange

SupercellSpec
Specifies a supercell.

Threading
A high-level control of how multiple cores are used.

UnfoldBands

Constants

MAX_VERSION

Traits

FailMessage

YamlRead
Provides an alternative to serde_yaml::from_reader where all of the expensive codegen has already been performed in this crate.

Type Definitions

CgStopCondition

Parameters