Struct rsp2_tasks_config::Phonons

pub struct Phonons {
    pub symmetry_tolerance: Option<f64>,
    pub displacement_distance: Option<f64>,
    pub analytic_hessian: bool,
    pub disp_finder: PhononDispFinder,
    pub eigensolver: PhononEigensolver,
    pub sum_rule: Option<PhononSumRule>,
    pub supercell: SupercellSpec,
}

Fields
symmetry_tolerance: Option<f64>

Cartesian distance threshold for determining if two sites are equivalent under a symmetry operator.

A value for this is required in most use cases. (the exception is when analytic-hessian: true)

If a value of 0 is used, symmetry will not be sought. This is necessary sometimes to work around limitations that prevent rsp2 from working on non-primitive cells.

displacement_distance: Option<f64>

How far atoms are displaced when numerically computing the force constants.

A value for this is required in most use cases. (the exception is when analytic-hessian: true)

analytic_hessian: bool

Use an analytically-computed hessian for the force constants.

If true, symmetry_tolerance, and displacement_distance are allowed to be null. Currently, symmetry will not be imposed in any form when this is true.

The vast majority of potentials do NOT support this. (in fact, no complete potential currently supports this; only the rust implementation of KCZ’s interlayer potential, without the contributions from REBO)

disp_finder: PhononDispFinder

How displacements are generated.

A value for this is required in most use cases. (the exception is when analytic-hessian: true)

eigensolver: PhononEigensolversum_rule: Option<PhononSumRule>

Method of imposing an acoustic sum rule on the force constants.

If not provided, no sum rule is imposed.

supercell: SupercellSpec

Supercell used for force constants.

Ideally, this should be large enough for the following to be true:

Given an atom with index p in the primitive cell…

… and an atom with index s in the supercell…

… there must be at most one image of s under the supercell lattice whose position affects the total force acting upon p.

The role of the supercell is twofold:

To ensure that multiple, distinct force terms are computed when a primitive atom p interacts with multiple images of a primitive atom q under the primitive lattice. (so that each interaction may receive a different phase factor when computing the dynamical matrix at nonzero Q points; strictly speaking, this concern does not apply to rsp2 as we only compute at the gamma point)

To isolate the effects of displacing one atom. For most potentials, the implementation will simply construct the supercell, displace the p site in that cell, and compute the new forces. This means that all images of p under the supercell will also move.

Thanks to the second point, even gamma point requires a supercell for sufficiently small structures. For instance, in REBO computed on the primitive unit cell of graphite, displacing one atom will also displace the atoms two bonds away, which would have an undesirable impact on the bond angle terms.