Struct rsp2_tasks_config::Settings

pub struct Settings {

    pub version: Option<u32>,
    pub threading: Threading,
    pub potential: ValidatedPotential,
    pub scale_ranges: ScaleRanges,
    pub parameters: Option<Parameters>,
    pub lattice_relax_22: Option<LatticeRelax>,
    pub acoustic_search: AcousticSearch,
    pub cg: Cg,
    pub phonons: Option<Phonons>,
    pub ev_chase: EigenvectorChase,
    pub layer_search: Option<LayerSearch>,
    pub bond_radius: Option<f64>,
    pub layer_gamma_threshold: f64,
    pub masses: Option<Masses>,
    pub ev_loop: EvLoop,
    pub unfold_bands: Option<UnfoldBands>,
    pub animate: Option<Animate>,
    pub snapshot: Snapshot,
    pub _deprecated_lammps_settings: DeprecatedLammpsSettings,
    pub lammps: Lammps,

}

Raw deserialized form of settings.

You shouldn’t deserialize this type directly; deserialize ValidatedSettings instead, so that additional validation and filling of defaults can be performed. (e.g. incompatible settings, or options whose defaults depend on others)

Fields
version: Option<u32>

Identifies the version of the settings that this file uses.

rsp2 increments the max supported version number when breaking changes are made to config files. Old config files with an older version continue to use the old behavior.

If not specified, assumes a value of 1.

threading: Threadingpotential: ValidatedPotential

Specifies the potential to be used.

See PotentialKind for the list of possibilities.

scale_ranges: ScaleRanges

Used to optimize lattice parameters prior to relaxation.

See the type for documentation.

parameters: Option<Parameters>

Names of parameters, using the same letter for parameters that should scale together.

This is used to enable optimization of lattice vector lengths during CG, if the potential supports it. (optimization of cell angles is not supported)

Note that the same letter does not mean that the lengths must be equal; it simply means that their ratio of lengths will be preserved. Use null (or equivalently ~) for parameters that should not be scaled. (e.g. vacuum separation)

Example:

# e.g. graphite
parameters: [a, a, c]

# e.g. chain along z
parameters: [~, ~, c]

lattice_relax_22: Option<LatticeRelax>

Provide relaxation of the top left 2x2 lattice submatrix.

If supplied (an empty mapping will do), parameters will be ignored and all four elements of the 2x2 submatrix will be optimized.

This is a temporary hack until a more natural design for how to write config for this can be designed.

Example:

lattice-relax-22: {}

acoustic_search: AcousticSearch

See the type for documentation.

cg: Cg

See the type for documentation.

phonons: Option<Phonons>

See the type for documentation.

ev_chase: EigenvectorChase

See the type for documentation.

layer_search: Option<LayerSearch>

None disables layer search. (layer_search is also ignored if layers.yaml is provided)

bond_radius: Option<f64>

None disables bond graph.

Sometimes this is ignored if a bond graph needs to be generated on the spot.

layer_gamma_threshold: f64masses: Option<Masses>

See the type for documentation.

ev_loop: EvLoop

See the type for documentation.

unfold_bands: Option<UnfoldBands>

None disables band unfolding.

animate: Option<Animate>

None disables animations.

snapshot: Snapshot

See the type for documentation.

_deprecated_lammps_settings: DeprecatedLammpsSettingslammps: Lammps

See the type for documentation.

Implementations

impl Settings

pub fn validate(self) -> Result<ValidatedSettings, Error>